Update README.md
This project consists of a dictionary developed in ReactJS with a list of English words, in which the user types the desired word and it shows the meanings and pronunciation. It is possible to listen to an audio with the pronunciation of each word.

In this project, the Redux Toolkit, Redux Thunk, the Axios library were used to consume the Free Dictionary API, react-icons for the icons, Tailwind CSS was used for the styles. The organization of this project is similar to the shopping cart system and the book list previously developed, but instead of using Sass I am using Tailwind CSS